+++ test/server/test_device_handler.py
@@ -17,7 +17,7 @@ from dataclasses import dataclass
 
 import scrutiny.server.datalogging.definitions.device as device_datalogging
 from scrutiny.server.timebase import server_timebase
-from scrutiny.server.device.emulated_device import UnitTestEmulatedDevice
+from scrutiny.server.device.emulated_device import EmulatedDevice
 from scrutiny.server.device.device_handler import DeviceHandler
 from scrutiny.server.device.links.dummy_link import DummyLink
 from scrutiny.server.datastore.datastore import Datastore
@@ -33,6 +33,29 @@ from test import ScrutinyUnitTest, logger
 from scrutiny.tools.typing import *
 
 no_callback: UpdateTargetRequestCallback = lambda *args, **kwargs: None
+
+
+class UnitTestEmulatedDevice(EmulatedDevice):
+    def __init__(self, *args:Any, **kwargs:Any) -> None:
+        super().__init__(*args, **kwargs)
+        
+        self.configure_rpvs({
+            0x1000: {'definition': RuntimePublishedValue(id=0x1000, datatype=EmbeddedDataType.float64), 'value': 0.0},
+            0x1001: {'definition': RuntimePublishedValue(id=0x1001, datatype=EmbeddedDataType.float32), 'value': 3.1415926},
+            0x1002: {'definition': RuntimePublishedValue(id=0x1002, datatype=EmbeddedDataType.uint16), 'value': 0x1234},
+            0x1003: {'definition': RuntimePublishedValue(id=0x1003, datatype=EmbeddedDataType.sint8), 'value': -65},
+            0x1004: {'definition': RuntimePublishedValue(id=0x1004, datatype=EmbeddedDataType.boolean), 'value': True}
+        })
+
+        self.configure_loops([
+            FixedFreqLoop(1000, name='1KHz'),
+            FixedFreqLoop(10000, name='10KHz'),
+            VariableFreqLoop(name='Variable Freq 1'),
+            VariableFreqLoop(name='Idle Loop', support_datalogging=False)
+        ])
+
+        self.configure_supported_features(memory_write=True, datalogging=True, user_command=True, _64bits=False)
+        self.configure_datalogger(256)
 
 
 def d2f(d):
